feat: program rules refactor
