+++ src/domain/usecases/ApplyInitialRulesToSurveyUseCase.ts
@@ -1,4 +1,4 @@
-import { Questionnaire, QuestionnarieM } from "../entities/Questionnaire";
+import { Questionnaire, QuestionnarieM } from "../entities/Questionnaire/Questionnaire";
 
 export class ApplyInitialRulesToSurveyUseCase {
     public execute(questionnaire: Questionnaire): Questionnaire {
